crossdock: Extract common CreateRPC function
