+++ crossdock/client/echo/behavior.go
@@ -21,17 +21,8 @@
 package echo
 
 import (
-	"fmt"
-	"net/http"
-
-	"github.com/yarpc/yarpc-go"
 	"github.com/yarpc/yarpc-go/crossdock/client/behavior"
 	"github.com/yarpc/yarpc-go/crossdock/client/params"
-	"github.com/yarpc/yarpc-go/transport"
-	ht "github.com/yarpc/yarpc-go/transport/http"
-	tch "github.com/yarpc/yarpc-go/transport/tchannel"
-
-	"github.com/uber/tchannel-go"
 )
 
 // echoEntry is an entry emitted by the echo behaviors.
@@ -70,38 +61,4 @@ func createEchoSink(encoding string, s behavior.Sink, p behavior.Params) behavio
 		Encoding:  encoding,
 		Server:    p.Param(params.Server),
 	}
-}
-
-// createRPC creates an RPC from the given parameters or fails the whole
-// behavior.
-func createRPC(s behavior.Sink, p behavior.Params) yarpc.RPC {
-	fatals := behavior.Fatals(s)
-
-	server := p.Param(params.Server)
-	fatals.NotEmpty(server, "server is required")
-
-	var outbound transport.Outbound
-	trans := p.Param(params.Transport)
-	switch trans {
-	case "http":
-		// Go HTTP servers have keep-alive enabled by default. If we re-use
-		// HTTP clients, the same connection will be used to make requests.
-		// This is undesirable during tests because we want to isolate the
-		// different test requests. Additionally, keep-alive causes the test
-		// server to continue listening on the existing connection for some
-		// time after we close the listener.
-		cl := &http.Client{Transport: new(http.Transport)}
-		outbound = ht.NewOutboundWithClient(fmt.Sprintf("http://%s:8081", server), cl)
-	case "tchannel":
-		ch, err := tchannel.NewChannel("client", nil)
-		fatals.NoError(err, "couldn't create tchannel")
-		outbound = tch.NewOutbound(ch, tch.HostPort(server+":8082"))
-	default:
-		fatals.Fail("", "unknown transport %q", trans)
-	}
-
-	return yarpc.New(yarpc.Config{
-		Name:      "client",
-		Outbounds: transport.Outbounds{"yarpc-test": outbound},
-	})
 }
