+++ crossdock/client/echo/behavior_test.go
@@ -95,52 +95,3 @@ func TestEchoSink(t *testing.T) {
 		}
 	}
 }
-
-func TestCreateRPC(t *testing.T) {
-	tests := []struct {
-		p      behavior.Params
-		errOut string
-	}{
-		{
-			behavior.ParamsFromMap{"server": "localhost"},
-			`unknown transport ""`,
-		},
-		{
-			behavior.ParamsFromMap{"transport": "http"},
-			"server is required",
-		},
-		{
-			behavior.ParamsFromMap{"server": "localhost", "transport": "foo"},
-			`unknown transport "foo"`,
-		},
-		{
-			p: behavior.ParamsFromMap{
-				"server":    "localhost",
-				"transport": "http",
-			},
-		},
-		{
-			p: behavior.ParamsFromMap{
-				"server":    "localhost",
-				"transport": "tchannel",
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		entries := behavior.Run(func(s behavior.Sink) {
-			rpc := createRPC(s, tt.p)
-
-			// should get here only if the request succeeded
-			ch := rpc.Channel("yarpc-test")
-			assert.Equal(t, "client", ch.Caller)
-			assert.Equal(t, "yarpc-test", ch.Service)
-		})
-
-		if tt.errOut != "" && assert.Len(t, entries, 1) {
-			e := entries[0].(behavior.Entry)
-			assert.Equal(t, behavior.Failed, e.Status)
-			assert.Contains(t, e.Output, tt.errOut)
-		}
-	}
-}
